commotalkie-12: Create a driver for EByte * Fix an argument type

+++ include/application/SubscriberBuilder.h
@@ -14,7 +14,7 @@ void SubscriberBuilder_Create();
 void SubscriberBuilder_SetSalt(const char *salt);
 
 void SubscriberBuilder_SetListenCallback(
-    int (*)(const char *, const char *, const int));
+    int (*)(const char *, const char *, const unsigned long));
 
 void SubscriberBuilder_SetTimeService(unsigned long (*)(void));
