+++ test/SubscriberBuilderTest.cpp
@@ -21,20 +21,20 @@ static unsigned long nothing_until_zero;
 
 // -----------------------------------------------------------------------------
 
-int stub_message_fn(const char *address, const char *content, const int size) {
+int stub_message_fn(const char *address, const char *content, const unsigned long size) {
   pull_fn_spy.calledCount++;
   Message message;
   MessageFormatter_Pack("0123456789AB", &message);
   MessageValidator_Sign(&message);
   memcpy((void *)content, (void *)&message, MESSAGE_LENGTH);
-  return size;
+  return (int)size;
 }
 
 unsigned long fake_epoch_ms_fn() {
   return 100;
 }
 
-int stub_pull_nothing_yet_fn(const char* address, const char* content, const int size) {
+int stub_pull_nothing_yet_fn(const char* address, const char* content, const unsigned long size) {
   if (--nothing_until_zero) {
     pull_fn_spy.calledCount++;
     return 0;
